trying osx again

+++ recipe/build.sh
@@ -19,34 +19,60 @@ export GMP_LIB_DIRS=$PREFIX/lib
 # Install shim scripts to ensure that certain flags are always passed to the compiler/linker
 #
 
-echo "#!/bin/bash" > $CC-shim
-echo "set -e -o pipefail -x " >> $CC-shim
-echo "$CC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CC-shim
-chmod u+x $CC-shim
-export CC=$CC-shim
+if [ -n "${CC:+1}" ]; then
+    echo "#!/bin/bash" > $CC-shim
+    echo "set -e -o pipefail -x " >> $CC-shim
+    echo "$CC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CC-shim
+    chmod u+x $CC-shim
+    export CC=$CC-shim
+fi
 
-echo "#!/bin/bash" > $GCC-shim
-echo "set -e -o pipefail -x " >> $GCC-shim
-echo "$GCC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GCC-shim
-chmod u+x $GCC-shim
-export GCC=$GCC-shim
+if [ -n "${CXX:+1}" ]; then
+    echo "#!/bin/bash" > $CXX-shim
+    echo "set -e -o pipefail -x " >> $CXX-shim
+    echo "$CXX -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $CXX-shim
+    chmod u+x $CXX-shim
+    export CXX=$CXX-shim
+fi
 
-echo "#!/bin/bash" > $LD-shim
-echo "set -e -o pipefail -x " >> $LD-shim
-echo "$LD -L$PREFIX/lib \"\$@\"" >> $LD-shim
-chmod u+x $LD-shim
-export LD=$LD-shim
+if [ -n "${GCC:+1}" ]; then
+    echo "#!/bin/bash" > $GCC-shim
+    echo "set -e -o pipefail -x " >> $GCC-shim
+    echo "$GCC -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GCC-shim
+    chmod u+x $GCC-shim
+    export GCC=$GCC-shim
+fi
 
-echo "#!/bin/bash" > ${LD}.gold
-echo "set -e -o pipefail -x " >> ${LD}.gold
-echo "$LD_GOLD -L$PREFIX/lib \"\$@\"" >> ${LD}.gold
-chmod u+x ${LD}.gold
-export LD_GOLD=${LD}.gold
+if [ -n "${GXX:+1}" ]; then
+    echo "#!/bin/bash" > $GXX-shim
+    echo "set -e -o pipefail -x " >> $GXX-shim
+    echo "$GXX -I$PREFIX/include -L$PREFIX/lib -pthread -fPIC \"\$@\"" >> $GXX-shim
+    chmod u+x $GXX-shim
+    export GXX=$GXX-shim
+fi
+
+if [ -n "${LD:+1}" ]; then
+    echo "#!/bin/bash" > $LD-shim
+    echo "set -e -o pipefail -x " >> $LD-shim
+    echo "$LD -L$PREFIX/lib \"\$@\"" >> $LD-shim
+    chmod u+x $LD-shim
+    export LD=$LD-shim
+fi
+
+if [ -n "${LD_GOLD:+1}" ]; then
+    echo "#!/bin/bash" > ${LD}.gold
+    echo "set -e -o pipefail -x " >> ${LD}.gold
+    echo "$LD_GOLD -L$PREFIX/lib \"\$@\"" >> ${LD}.gold
+    chmod u+x ${LD}.gold
+    export LD_GOLD=${LD}.gold
+fi
 
 HOST_LIBPTHREAD="${BUILD_PREFIX}/${HOST}/sysroot/usr/lib/libpthread.so"
 
-rm ${HOST_LIBPTHREAD}
-ln -s /lib64/libpthread.so.0 ${HOST_LIBPTHREAD}
+if [[ -f "${HOST_LIBPTHREAD}" ]]; then
+    rm ${HOST_LIBPTHREAD}
+    ln -s /lib64/libpthread.so.0 ${HOST_LIBPTHREAD}
+fi
 
 #######################################################################################################
 # Install bootstrap ghc
